WIP: switch to ini-style config
This commit is a breaking change but will simplify adding config options
in the future since a user will only have to add the settings unique to
them to their config file instead of having to have ALL settings in
their file. This will mean that the program will operate so long as
there an api key is defined in the LOCAL_SETTINGS section of config.ini.